--- src/app/components/shared/menu/external-link/external-link.component.scss
@@ -1,20 +1,0 @@
-.disabled {
-  cursor: not-allowed;
-  opacity: 0.80;
-}
-
-.nav-link {
-  color: black;
-  padding: 0.5rem 0.25rem 0.5rem 0.75rem;
-
-  &:hover,
-  &:focus {
-    background-color: var(--theme-highlight-light);
-  }
-
-  .icon {
-    display: inline-flex;
-    justify-content: center;
-    width: 1.3rem;
-  }
-}
